document default compile options. Support C lang

+++ tests/libcstr/tests/test.c
@@ -1,7 +1,8 @@
 #include <libcstr/cstr.h>
 
-int main() {
+int main()
+{
   Cstr str = cstr_create("a string");
   cstr_destroy(&str);
-  return str.size;
+  return (int)(str.size);
 }
